requirements.txt
uiautomation
pynput
Pillow
xhtml2pdf
Jinja2
